complete the docker provider

+++ manager-init.tpl
@@ -5,12 +5,18 @@ set -o xtrace
 
 {{ source "default.ikt" }}
 {{ source "file:///infrakit/env.ikt" }}
-{{ include "install-docker.sh" }}
-{{ source "attach-ebs-volume.sh" }}
 {{ source "provider.sh" }}
+{{ source "install-docker.sh" }}
+{{ source "attach-ebs-volume.sh" }}
+
+# if we're already in a Docker container, either it's DinD and Docker is already installed
+# or it's a non DinD Docker container and we can't easily install Docker
+if [ "x$provider" != "xdocker" ]; then
+  _install_docker
+  systemctl stop docker.service
+fi
 
 # Use an EBS volume for the devicemapper
-systemctl stop docker.service
 if [ "x$provider" = "xaws" ]; then
   rm -rf /var/lib/docker
   _attach_ebs_volume /dev/sdn /var/lib/docker "Docker AUFS" {{ ref "/docker/aufs/size" }}
@@ -23,12 +29,21 @@ cat << EOF > /etc/docker/daemon.json
 }
 EOF
 
-systemctl start docker.service
-sleep 2
+if [ "x$provider" != "xdocker" ]; then
+  systemctl start docker.service
+  sleep 2
+fi
 
 {{ if ref "/certificate/ca/service" }}{{ include "request-certificate.sh" }}{{ end }}
 
+# INSTANCE_LOGICAL_ID can be an IP of a hostname, we need an IP
+IP="{{ INSTANCE_LOGICAL_ID }}"
+if [ ! $(echo "$IP" | egrep -q "([0-9.]+){4}") ]; then
+  IP=$(nslookup {{ INSTANCE_LOGICAL_ID }}  2>/dev/null | awk '$1 == "Address" {print $3}' | tail -1)
+fi
+
 {{ if and ( eq INSTANCE_LOGICAL_ID SPEC.SwarmJoinIP ) (not SWARM_INITIALIZED) }}
+if [ "x$provider" != "xdocker" ]; then
   mkdir -p /etc/systemd/system/docker.service.d
   cat > /etc/systemd/system/docker.service.d/docker.conf <<EOF
 [Service]
@@ -39,9 +54,10 @@ EOF
   # Restart Docker to let port listening take effect.
   systemctl daemon-reload
   systemctl restart docker.service
+fi
 
   {{/* The first node of the special allocations will initialize the swarm. */}}
-  docker swarm init --advertise-addr {{ INSTANCE_LOGICAL_ID }}
+  docker swarm init --advertise-addr $IP
 
 {{ else }}
